requirements.txt
streamlit==1.23.1
python-dotenv==0.21.0
pandas==1.3.5
folium==0.12.1
streamlit-folium==0.12.0
pyttsx3==2.90
plotly==5.11.0
